+++ tests/utils/mod.rs
@@ -1,6 +1,6 @@
 use game_of_life::renderer::CellPoint;
 
-pub fn convert_coords(input: Vec<(i64, i64)>) -> Vec<CellPoint> {
+pub fn convert_coords(input: Vec<(i128, i128)>) -> Vec<CellPoint> {
     input.into_iter().map(CellPoint::from_tuple).collect()
 }
